% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caldelta.R
\name{caldelta}
\alias{caldelta}
\title{estimate the threshold for power enhancement component.}
\usage{
caldelta(X, G, B = 2000, seed = 1, simple = FALSE)
}
\arguments{
\item{X}{A matrix of adjusting covariates.}

\item{G}{A matrix of genetic variants.}

\item{B}{The number of the replicates.}

\item{seed}{The random seed.}

\item{simple}{logical. If true, use the sampling method to estimate the threshold.}
}
\value{
returns the threshold for the power enhancement.
}
\description{
estimate the threshold for power enhancement component.
}
