% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdtest.R
\name{hdtest}
\alias{hdtest}
\title{High-dimensional association test with one response}
\usage{
hdtest(
  Y,
  X,
  G,
  sigma2 = NULL,
  method = "high",
  ape = TRUE,
  delta = NULL,
  boncor = TRUE,
  edgecor = FALSE,
  ...
)
}
\arguments{
\item{Y}{The response.}

\item{X}{A matrix of adjusting covariates.}

\item{G}{A matrix of genetic variants.}

\item{sigma2}{The estimated error variance. If null, it will be
estimated by refitted cross validation.}

\item{method}{the method to test the association. The default method 'high' accounts
the accumulated estimation error under high dimensions.}

\item{ape}{a logical value indicating whether the power enhancement component is used.}

\item{delta}{the threshold for the power enhancement.}

\item{boncor}{a logical value indicating whether the p-value from the Bonferroni
correction should be reported.}

\item{edgecor}{a logical value indicating whether the Edgeworth expansion is used
for extreme significance levels.}
}
\value{
A list containing the test statistics and the associated p-values.
\item{TH}{the observed TH test statistic}
\item{ph}{the p-value of the TH test}
\item{sd1}{the standard deviation of the test statistic}
\item{Tpe}{the observed test statistic with the power enhancement}
\item{phe}{the p-value of the test with the power enhancement}
\item{pb}{the p-value of the marginal models after Bonferroni correction}
\item{ph_edge}{the p-value of the TH test using Edgeworth expansion}
\item{phe_edge}{the p-value of the test with the power enhancement component using Edgeworth expansion}
\item{diff}{difference of p^-1||Sigma||_F^2 - p/n to verify a condition in the paper}
}
\description{
Test the association between the outcome Y and genetic variants G under high dimensions.
}
