% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvtest.R
\name{mvtest}
\alias{mvtest}
\title{High-dimensional association test with multivariate responses}
\usage{
mvtest(Y, X, G, Sigma2 = NULL)
}
\arguments{
\item{Y}{A matrix of multivariate outcomes.}

\item{X}{A matrix of adjusting covariates.}

\item{G}{A matrix of genetic variants.}

\item{Sigma2}{The estimated covariance matrix of Y. If null, it will be
estimated by residuals from the marginal models.}
}
\value{
A list containing the test statistics and the associated p-values.
\item{T1}{the observed T1 test statistic}
\item{pval_T1}{the p-value of the T1 test}
\item{T2}{the observed T2 test statistic}
\item{pval_T2}{the p-value of the T2 test}
}
\description{
Test the association between multivariate outcomes Y and genetic variants G
}
