% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vrcv.R
\name{vrcv}
\alias{vrcv}
\title{estimate the error variance in linear regression using refitted cross validation.}
\usage{
vrcv(Y, X, Num = 10, sis_prop = 0.5, seed = NULL)
}
\arguments{
\item{Y}{The response vector.}

\item{X}{A matrix of adjusting covariates.}

\item{Num}{The number of the replicates.}

\item{sis_prop}{The proportion in SIS to select relevant features.}

\item{seed}{The random seed.}
}
\value{
returns the estimated error variance.
}
\description{
estimate the error variance in linear regression using refitted cross validation.
}
